Remove now redundant specs
This is testing framework behaviour. We can be sure these values are
type-cast for the configured database adapter.

Removing this silences Rails 5 warning messages around the coercion
of truthy values to booleans.

We seemed to be dependent on anything outside of the usually coercable
booleans (`'true'`, `'false'`, `'1'`, `'0'`, `'y'`, `'n'` etc) to be
converted to `false` which would be surprising first off, and is not
actually behaviour we depend on inside of our own application code.