Fix positioning of numeric error messages
If a field is tagged as `js-number-only` and a
non-numerical value is input, the error class will
appear on the screen and then fade out.

Changed the code so the error text appears next to
the input box that generated the error with support
for multiple error locations on a single form.