Open PDF in separate window for Postal delivery
To enable TPAS to print and send summary documents,
if postal option is selected submitting the form
from the preview screen will save the document, open
the PDF in a separate window and then navigate to
the done screen.

This also fixes a bug where refreshing the `done` screen
created a duplicate summary document record.