Poll for notify delivery statuses
We need to poll for notify email delivery statuses to ensure we remain
deliverable at all times. Unfortunately notify does not supply webhooks
so we have to poll for the statuses instead.

A summary will be included for polling when it's less than two days old,
has a notify supplied ID and its `notify_completed_at` timestamp is
absent. When notify deems itself 'done' with our notification whether
due to a permanent failure or delivery it will supply us with the
`completed_at` timestamp that we use to mark the summary for exclusion
next time around.