Remove preview page and replace with confirmation page
Feedback suggested that previewing the document before
printing/saving was an overkill and wasn't being checked
over as the document is often so long.

This replaces the preview of the PDF with a confirmation
page which highlights key information about the customer
and brings into line with the confirmation screen in
TAP when you book an appointment.