Correct issue with disabled/readonly field
When the field is marked with the `disabled` attribute it does not POST
a value, thus would be preselected as `standard` for the
`appointment_type`. This means some customers were missing out on vital
information pertaining to the pensionable age.

I also scoped the test helper methods as they were clobbering eachother
in the global namespace.